+++ src/extension.ts
@@ -1,15 +1,15 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import { ContentsTreeView, Content } from "./contentsTreeView";
+import { CContensViewer } from './ContensViewer';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
 
 	// const contentsProvider = new ContentsTreeView();
-	// vscode.window.registerTreeDataProvider('ContentsViewer-view', contentsProvider);
-	const treeview = new ContentsTreeView(context);
+	// vscode.window.registerTreeDataProvider('ContentsViewer-view', contentsProvider)
+	const contentsViewer = new CContensViewer(context);
 
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
@@ -24,26 +24,34 @@ export function activate(context: vscode.ExtensionContext) {
 
 	vscode.commands.registerCommand('ContentsViewer.refresh', () => {
 		// The code you place here will be executed every time your command is executed
-		treeview.refresh();
+		contentsViewer.refresh();
 		const editor = vscode.window.activeTextEditor?.selection;
 		if (editor) {
-			treeview.highlightItemByCurLineNum();
+			contentsViewer.highlightItemByCurLineNum();
 		}
 	});
 
 	vscode.commands.registerCommand('ContentsViewer.headIndex', () => {
-		treeview.headIndex();
+		contentsViewer.headIndex();
 		const editor = vscode.window.activeTextEditor?.selection;
 		if (editor) {
-			treeview.highlightItemByCurLineNum();
+			contentsViewer.highlightItemByCurLineNum();
 		}
 	});
 
 	vscode.commands.registerCommand('ContentsViewer.specificIndex', () => {
-		treeview.specificIndex();
+		contentsViewer.specificIndex();
 		const editor = vscode.window.activeTextEditor?.selection;
 		if (editor) {
-			treeview.highlightItemByCurLineNum();
+			contentsViewer.highlightItemByCurLineNum();
+		}
+	});
+
+	vscode.commands.registerCommand('ContentsViewer.numberContents', () => {
+		contentsViewer.numberContents();
+		const editor = vscode.window.activeTextEditor?.selection;
+		if (editor) {
+			contentsViewer.highlightItemByCurLineNum();
 		}
 	});