+++ CHANGELOG.md
@@ -12,3 +12,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Click to goto the line of the content
 - Support to hightlight the item of contents' treeview corresponding to the current row
 - head index and specific index
+- number contents